FITRUN-177 refactor: 불필요한 파일 삭제 및 패키지 통합
