+++ src/main/kotlin/com/yapp/yapp/audio/domain/AudioServletHandler.kt
@@ -42,7 +42,7 @@ object AudioServletHandler {
     fun createAudioHeader(): HttpHeaders {
         return HttpHeaders().apply {
             contentType = MediaType.parseMediaType("audio/mpeg")
-            add(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=\"speech.mp3\"")
+            add(HttpHeaders.CONTENT_DISPOSITION, "inline")
         }
     }
 }
